Test refactor. The getsgeos tests are now in their own file that respect build restrictions that reduce compile times. PTE tests still built under normal conditions.
